Refactor: Overhaul Gastos feature and add navigation
This commit introduces a major refactoring of the Gastos (Expenses) feature and sets up the initial navigation structure.

Key changes include:
- **API Layer:**
    - Added an endpoint to find a `Gasto` by its ID (`findGasto`).
    - Implemented an `OkHttpClient` with an `HttpLoggingInterceptor` for improved debugging of network requests.
    - Updated `GastoRequestDto` to remove `gastoId` and changed `itbis` and `monto` types from `Int` to `Double`.

- **Repository & Data Source:**
    - Implemented `findGasto` functionality in the repository and remote data source.
    - Refactored `getGastos` in the repository to directly return a list of domain `Gasto` objects.

- **Domain Layer & Use Cases:**
    - The `Gasto` domain model now has default values, and `itbis` and `monto` are `Double`.
    - Introduced `FindGastoUseCase` and added validation logic to `CrearGastoUseCase`.
    - Created `GastoUseCases` to group all related use cases.
    - Simplified `GetGastosUseCase` implementation.

- **Dependency Injection:**
    - Restored and configured `DomainModule` and `RepositoryModule` for Hilt.
    - Updated `ApiModule` to provide `Retrofit` and `OkHttpClient`.

- **UI & App Structure:**
    - `MainActivity` now uses a `NavHost` for navigation.
    - Added internet permissions to `AndroidManifest.xml`.
    - Added new dependencies for testing.

+++ app/src/main/java/edu/ucne/enmanuel_gomez_ap2_p2/data/gastos/remote/dto/GastoResponseDto.kt
@@ -5,6 +5,6 @@ data class GastoResponseDto(
     val fecha: String,
     val suplidor: String,
     val ncf: String,
-    val itbis: Int,
-    val monto: Int,
+    val itbis: Double,
+    val monto: Double,
 )
